% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_equal.R
\name{all_equal}
\alias{all_equal}
\title{Check vector items for equality}
\usage{
all_equal(vec)
}
\arguments{
\item{vec}{a vector of any type that supports checking for equality (e.g. character, numeric, etc)}
}
\value{
logical
}
\description{
Tests whether every element in a vector is equal
}
\examples{
all_equal(rep("apple",10))
all_equal(rep(c("apple","pear"),10))

}
