% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_cross_validation.R
\name{app_minus_test_thresh_weighted_sum}
\alias{app_minus_test_thresh_weighted_sum}
\title{Weight feature importance sum on the basis of overfitting}
\usage{
app_minus_test_thresh_weighted_sum(
  k_cv_res,
  feature_importance_mat,
  k,
  penalty = 5
)
}
\arguments{
\item{k_cv_res}{result from run_cross_validation (called with condensed_output = TRUE)}

\item{feature_importance_mat}{feature importance values from each kTSCR iteration derived from how many times a feature was seen in a sibling}

\item{k}{k number of cross validation iterations, passed from run_cross_validation function}

\item{penalty}{a penalty used in the weighting. A (potential) hyperparameter. Default is 5.}
}
\value{
numeric vector of weighted feature importance sums
}
\description{
When calculating feature importance by tallying the number of times a feature is in a sibling within and then across k-cv iterations, weight the tally by the overfitting seen in the iteration, as measured by app_cor - test_cor
}
\examples{

}
