% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_top_features.R
\name{choose_top_features}
\alias{choose_top_features}
\title{Choose top features}
\usage{
choose_top_features(vec)
}
\arguments{
\item{vec}{a numerical vector of features importance scores. Vector should be named with feature names}
}
\value{
a vector of top_features
}
\description{
Use kmeans clustering to separate features into priority groups based on feature importance scores calculated during n fold k-cv iterations
}
\examples{


}
