% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_siblings.R
\name{convert_sibling_names_to_indices}
\alias{convert_sibling_names_to_indices}
\title{Convert sibling names to indices}
\usage{
convert_sibling_names_to_indices(X, siblings)
}
\arguments{
\item{X}{input matrix}

\item{siblings}{character vector (as a matrix) of feature names}
}
\value{
character vector (as a matrix) of feature indices
}
\description{
Convert sibling names to indices
}
