% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_siblings.R
\name{elder_weighted_corr_score}
\alias{elder_weighted_corr_score}
\title{Calculate weighted corr score}
\usage{
elder_weighted_corr_score(corr, max_score = 100)
}
\arguments{
\item{corr}{a float between 0 and 1 representing the absolute value of a correlation coefficient (r)}

\item{max_score}{the maximum weight given to a correlation. Given that this weighted scoring can produce exponentially increasing weighted scores, setting this to 100 (representing a correlation of ~0.99) is reasonable.}
}
\value{
a float between 0 and 100, that represents a mapping of corr to a weighted score that overweights higher correlation values
}
\description{
Calculates a weighted correlation score that over weights higher correlations. When calling elders, this allows me to prioritize features that have are a part of a smaller number of highly correlative (with y) pairwise features as opposed to features that are a part of a larger number of relative less correlative pairwise features.
}
\examples{
corr <- 0.05
weighted_score <- elder_weighted_corr_score(corr)

}
