% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_cross_validation.R
\name{get_app_minus_test}
\alias{get_app_minus_test}
\title{Calculate apparent correlation minus test correlation}
\usage{
get_app_minus_test(k_cv_res, k)
}
\arguments{
\item{k_cv_res}{result from run_cross_validation (called with condensed_output = TRUE)}

\item{k}{k number of cross validation iterations, passed from run_cross_validation function}
}
\value{
a numeric vector with app_cor - test_cor
}
\description{
Calculate app_cor - test_cor as a measure of overfitting at each k-cv iteration
}
\examples{

}
