% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_siblings.R
\name{get_sibling_indices}
\alias{get_sibling_indices}
\title{Get indices of sibling features}
\usage{
get_sibling_indices(siblings)
}
\arguments{
\item{siblings}{a vector of sibling names (each of which is a character vector, i.e. string)}
}
\value{
returns siblings as a numeric vector of indices rather than the character vector version of said indices
}
\description{
Convert sibling feature names into indices from the original feature matrix X
}
\examples{
siblings <- c("V1_V3", "V20_V5", "V100_V110", "V101_V50")
sibling_indices <- get_sibling_indices(siblings)


}
