% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_siblings.R
\name{get_sorted_corrs_pairwise_features}
\alias{get_sorted_corrs_pairwise_features}
\title{Correlations between each pairwise feature and y}
\usage{
get_sorted_corrs_pairwise_features(pairwise_feature_mat, y)
}
\arguments{
\item{pairwise_feature_mat}{a matrix with scores from feature pair comparison (rows) for each sample (cols)}

\item{y}{vector with quantitative outcome variable}
}
\value{
a sorted list of pairwise features based on their pearson correlation with y
}
\description{
Correlations between each pairwise feature and y
}
\examples{
C <- 100  # represents samples
R <- 200 # represents features
y <- rnorm(C) # represents outcome variable
X <- matrix(rbeta(R*C, 2, 3), nrow = R)  # simulate data matrix
Is <- get_pairwise_rank_matrices(X)
pairwise_feature_mat <- make_feature_pair_score_matrix(Is)
sorted_corrs <- get_sorted_corrs_pairwise_features(pairwise_feature_mat, y)

}
