% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_feature_pair_score_matrix.R
\name{make_feature_pair_score_matrix}
\alias{make_feature_pair_score_matrix}
\title{Combine per sample feature pair score matrices into a single non-redundant matrix (i.e. keep Xi - Xj, but not Xj - Xi, and have scores for every sample in a single matrix)}
\usage{
make_feature_pair_score_matrix(Is)
}
\arguments{
\item{Is}{list of matrices, where each matrix represents a feature x feature score matrix for a single sample}
}
\value{
a matrix of feature pair score for each sample
}
\description{
Combine per sample feature pair score matrices into a single non-redundant matrix (i.e. keep Xi - Xj, but not Xj - Xi, and have scores for every sample in a single matrix)
}
\examples{
C <- 100  # represents samples
R <- 200 # represents features
X <- matrix(rbeta(R*C, 2, 3), nrow = R)  # simulate data matrix
Is <- get_pairwise_rank_matrices(X)
pairwise_feature_mat <- make_feature_pair_score_matrix(Is)

}
