% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict}
\alias{predict}
\title{Predict y from K}
\usage{
predict(y, K)
}
\arguments{
\item{y}{vector with quantitative outcome variable}

\item{K}{K score calculated by kTSCR procedure}
}
\value{
numeric vector of predicted values for y
}
\description{
Predict y from K
}
\examples{

}
