% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmse.R
\name{rmse}
\alias{rmse}
\title{Calculate Root Mean Squared Error}
\usage{
rmse(y, y_pred)
}
\arguments{
\item{y}{vector of numeric values, such as the outcome variable in regression}

\item{y_pred}{another numeric vector, such as y values predicted by a model}
}
\value{
numeric respresent the rmse
}
\description{
Calculate Root Mean Squared Error
}
\examples{

}
