% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_siblings.R
\name{split_sibling_indices}
\alias{split_sibling_indices}
\title{Split pairwise features into numeric indices}
\usage{
split_sibling_indices(siblings)
}
\arguments{
\item{siblings}{a vector of sibling names made of indices concatenated by underscore (is a string)}
}
\value{
returns siblings as a numeric vector of indices rather than the character vector version of said indices
}
\description{
Convert character vector of pairwise features (i.e. siblings)  into numeric matrix of indices from the original feature matrix X
}
\examples{
siblings <- c("1_3", "20_5", "100_110", "101_50")
sibling_indices <- split_sibling_indices(siblings)


}
