% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_siblings.R
\name{split_sibling_names}
\alias{split_sibling_names}
\title{Split pairwise features into character vector}
\usage{
split_sibling_names(siblings)
}
\arguments{
\item{siblings}{a vector of sibling names (each of which is a character vector, i.e. string)}
}
\value{
returns siblings as a character vector of split names reflecting pairs of independent features from the original matrix X
}
\description{
Convert character vector of concatenated pairwise features (i.e. siblings)  into a 'character vector' matrix of split names (i.e. original non-pairwise feature names)
}
\examples{
siblings <- c("V1_V3", "V20_V5", "V100_V110", "V101_V50")
sibling_indices <- split_sibling_names(siblings)


}
