% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_pairwise_feature_mat.R
\name{update_pairwise_feature_mat}
\alias{update_pairwise_feature_mat}
\title{Remove elder from pairwise feature matrix}
\usage{
update_pairwise_feature_mat(pairwise_feature_mat, elder)
}
\arguments{
\item{pairwise_feature_mat}{a matrix of feature pair score for each sample}

\item{elder}{character vector (i.e. string) representing the name of the elder feature}
}
\value{
returns pairwise feature matrix but without the columns that had pairwise features where one feature was the elder feature
}
\description{
Remove all columns from pairwise_feature_mat that have elder as one of the features in each pairwise feature
}
\examples{
C <- 100  # represents samples
R <- 200 # represents features
y <- stats::rnorm(C) # represents outcome variable
X <- matrix(rbeta(R*C, 2, 3), nrow = R)  # simulate data matrix
Is <- get_pairwise_rank_matrices(X)
pairwise_feature_mat <- make_feature_pair_score_matrix(Is)
sorted_corrs <- get_sorted_corrs_pairwise_features(pairwise_feature_mat, y)
elder_corr <- get_elder(sorted_corrs)
elder <- names(elder_corr)
pairwise_feature_mat <- update_pairwise_feature_mat(pairwise_feature_mat, elder)

}
