% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_cross_validation.R
\name{condense_k_cv_output}
\alias{condense_k_cv_output}
\title{Condense k cross validation output}
\usage{
condense_k_cv_output(k_cv_res, k)
}
\arguments{
\item{k_cv_res}{k cross validation result object from function run_cross_validation()}

\item{k}{k number of cross validation iterations, passed from run_cross_validation function}
}
\value{
list of four vectors: app_cor, test_cor, elders, feature_importance
}
\description{
Condense and summarize output from k cross validation iterations. Specifically, this returns the union of elders across k-cv iterations, and the total per feature of how many sibling pairs it was in across all k-cv iterations
}
\examples{

}
